refact: move theme files to ds.atoms
