+++ app/src/main/java/br/com/mob1st/bet/MainActivity.kt
@@ -7,7 +7,7 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.tooling.preview.Preview
 import br.com.mob1st.bet.screens.MainScreen
-import br.com.mob1st.bet.ui.theme.BetTheme
+import br.com.mob1st.bet.core.ui.ds.atoms.BetTheme
 
 class MainActivity : ComponentActivity() {
