+++ app/src/main/java/br/com/mob1st/bet/core/ui/ds/atoms/Color.kt
@@ -1,4 +1,4 @@
-package br.com.mob1st.bet.ui.theme
+package br.com.mob1st.bet.core.ui.ds.atoms
 
 import androidx.compose.ui.graphics.Color
