+++ app/src/main/java/br/com/mob1st/bet/core/ui/ds/atoms/Type.kt
@@ -1,4 +1,4 @@
-package br.com.mob1st.bet.ui.theme
+package br.com.mob1st.bet.core.ui.ds.atoms
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
